refactor(latent_codecs): extract checkerboard helpers directly inside checkerboard.py
